Refactor accept / reject change buttons out of time change component
